# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/parivaar/public/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/v1/addresses": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "API to get addresses of user",
                "operationId": "indexad",
                "responses": {
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create address",
                "description": "This api will create new address.",
                "operationId": "storeaaa",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "address_type": {
                                        "type": "number",
                                        "example": "1:Billing,2:Shipping"
                                    },
                                    "is_default": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John wick"
                                    },
                                    "address1": {
                                        "type": "string",
                                        "example": "Wall Street 1"
                                    },
                                    "address2": {
                                        "type": "string",
                                        "example": "Near ATM"
                                    },
                                    "country": {
                                        "type": "number",
                                        "example": 103
                                    },
                                    "state": {
                                        "type": "number",
                                        "example": 21
                                    },
                                    "district": {
                                        "type": "number",
                                        "example": 21
                                    },
                                    "city": {
                                        "type": "number",
                                        "example": 21
                                    },
                                    "pincode": {
                                        "type": "number",
                                        "example": 741258
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}